Was made save changes method, upload changes methodand was made some refactoring
