+++ main_game.h
@@ -1,6 +1,8 @@
 #include "create_map.h"
 #include "print_map.h"
 #include "copy_map.h"
+#include "restore_changes.h"
 #include "save_changes.h"
+#include "release_memory.h"
 
-void game_start();
+void game_start(int fc, int mc, int ic, int rc);
