+++ makefile
@@ -1,17 +1,20 @@
 CC = gcc
 OPTIONS = -Wall -ansi -pedantic
-
 EXECUTABLE = rahan
-
 SOURCE = $(wildcard *.c)
-OBJECTS = $(SOURCE:.c=.o)
+OBJ_DIR = obj_dir
+OBJECTS = $(addprefix $(OBJ_DIR)/, $(notdir $(SOURCE:.c=.o)))
 
 all: $(EXECUTABLE)
 
 $(EXECUTABLE): $(OBJECTS)
-	$(CC) $(OPTIONS) $(OBJECTS) -o $(EXECUTABLE)
+	$(CC) $(OPTIONS) $^ -o $@
 
-%.o: %.c %.h
-	$(CC) $(OPTIONS) -MMD -c $<
+$(OBJ_DIR)/%.o: %.c
+	$(CC) $(OPTIONS) -MMD -c $< -o $@
 
--include *.d
+-include $(OBJ_DIR)/*.d
+
+.PHONY: clean
+clean:
+	rm -f $(EXECUTABLE) $(OBJ_DIR)/*.o $(OBJ_DIR)/*.d
