+++ restore_changes.h
@@ -1,0 +1,3 @@
+#include "types.h"
+
+Map* restore_map(int foodCount, int monsterCount, int itemsCount, int rocksCount);
