+++ save_changes.h
@@ -1,4 +1,3 @@
 #include "types.h"
 
-Map* restore_map(int foodCount, int monsterCount, int itemsCount, int rocksCount);
 void upload_changes(Map* m);
