--- test.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <limits.h>
-#include <time.h>
-#include "main_game.h"
-
-int main(void){
-    game_start();
-
-    stdprof_afficher();
-
-    return 0;
-}
