+++ txts/checkpoint.txt
@@ -1,0 +1,113 @@
+Food:3
+Enchanted Stardust Cake
+The wizards also refer to them as space cakes
+6
+*
+1
+6
+Eternal Emberberries
+A berry growing for eternities and counting
+24
+*
+1
+1
+Velvet Skyfruit
+Growing on the sky islands, a rare delicacy
+2
+*
+5
+0
+
+Rocks:4
+#
+2
+5
+#
+6
+6
+#
+4
+8
+#
+7
+1
+
+Monsters:3
+Grimreaper
+7
+17
+@
+7
+0
+Semy
+9
+4
+@
+1000
+9
+Nightshade
+-14
+12
+@
+1000
+4
+
+Items:4
+Soulforge Crown
+0
+5
+!
+1
+9
+Obsidian Eclipse Blade
+1
+3
+!
+1000
+4
+Abyssal Abysswalkers
+2
+4
+!
+1000
+7
+Starfall Sentinel
+0
+2
+!
+3
+2
+
+Hero:
+53
+38
+14
+1
+2
+2
+R
+8
+0
+Hero_Items:3
+null
+0
+0
+null
+Obsidian Eclipse Blade
+1
+3
+!
+Abyssal Abysswalkers
+2
+4
+!
+
+Map:
+3
+16
+10
+10
+0
+0
+0
+0
