+++ txts/itemnames.txt
