+++ txts/monsternames.txt
